test: migrate integration tests and test data; fix integration tests
