+++ ors-api/src/test/files/data/preprocessed_osm_data.pbf
