+++ ors-api/src/test/java/org/heigit/ors/apitests/centrality/ResultTest.java
@@ -1,9 +1,9 @@
 package org.heigit.ors.apitests.centrality;
 
 import io.restassured.response.Response;
-import org.heigit.ors.apitests.common.EndPointAnnotation;
 import org.heigit.ors.apitests.common.ServiceTest;
 import org.heigit.ors.apitests.common.VersionAnnotation;
+import org.heigit.ors.apitests.common.EndPointAnnotation;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
