+++ ors-api/src/test/java/org/heigit/ors/apitests/isochrones/fast/ParamsTest.java
@@ -11,13 +11,16 @@
  *  You should have received a copy of the GNU Lesser General Public License along with this library;
  *  if not, see <https://www.gnu.org/licenses/>.
  */
-package org.heigit.ors.apitests.isochrones;
+package org.heigit.ors.apitests.isochrones.fast;
 
+import org.hamcrest.Matchers;
 import org.heigit.ors.apitests.common.EndPointAnnotation;
 import org.heigit.ors.apitests.common.ServiceTest;
 import org.heigit.ors.apitests.common.VersionAnnotation;
+import org.heigit.ors.apitests.isochrones.IsochronesErrorCodes;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -34,8 +37,7 @@ class ParamsTest extends ServiceTest {
     public ParamsTest() {
         // Locations
         addParameter("preference", "fastest");
-        addParameter("cyclingProfile", "cycling-regular");
-        addParameter("carProfile", "driving-car");
+        addParameter("hgvProfile", "driving-hgv");
 
         JSONArray firstLocation = new JSONArray();
         firstLocation.put(8.681495);
@@ -102,7 +104,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -120,12 +122,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.MISSING_PARAMETER))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.MISSING_PARAMETER))
                 .statusCode(400);
     }
 
@@ -138,12 +140,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.UNKNOWN_PARAMETER))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.UNKNOWN_PARAMETER))
                 .statusCode(400);
     }
 
@@ -159,12 +161,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_FORMAT))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_FORMAT))
                 .statusCode(400);
     }
 
@@ -180,12 +182,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
 
     }
@@ -201,12 +203,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MINIMUM))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MINIMUM))
                 .statusCode(400);
     }
 
@@ -222,12 +224,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM))
                 .statusCode(400);
     }
 
@@ -244,12 +246,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
 
     }
@@ -265,12 +267,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
     }
 
@@ -284,12 +286,12 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
     }
 
@@ -306,26 +308,26 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
+                .post(getEndPointPath()+"/{profile}/geojson")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
 
         body.put("attributes", new JSONArray(Arrays.asList("blah", "reachfactor")));
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
+                .post(getEndPointPath()+"/{profile}/geojson")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
     }
 
@@ -344,13 +346,13 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
+                .post(getEndPointPath()+"/{profile}/geojson")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
     }
 
@@ -368,13 +370,13 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
+                .post(getEndPointPath()+"/{profile}/geojson")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE))
                 .statusCode(400);
     }
 
@@ -387,28 +389,29 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/blah")
+                .post(getEndPointPath()+"/{profile}/blah")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.UNSUPPORTED_EXPORT_FORMAT))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.UNSUPPORTED_EXPORT_FORMAT))
                 .statusCode(406);
 
         given()
                 .headers(jsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
+                .post(getEndPointPath()+"/{profile}/geojson")
                 .then()
                 .assertThat()
-                .body("error.code", is(IsochronesErrorCodes.UNSUPPORTED_EXPORT_FORMAT))
+                .body("error.code", Matchers.is(IsochronesErrorCodes.UNSUPPORTED_EXPORT_FORMAT))
                 .statusCode(406);
     }
 
     @Test
+    @Disabled("Currently no support for location type destination with fast isochrones")
     void testDestination() {
 
         JSONObject body = new JSONObject();
@@ -419,8 +422,9 @@ class ParamsTest extends ServiceTest {
         body.put("location_type", "destination");
 
         given()
-                .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .header("Accept", "application/geo+json")
+                .header("Content-Type", "application/json")
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -440,7 +444,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -461,7 +465,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -483,7 +487,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -506,7 +510,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -527,11 +531,11 @@ class ParamsTest extends ServiceTest {
         body.put("range", ranges);
         body.put("range_type", "time");
         body.put("interval", getParameter("interval_200"));
-        body.put("location_type", "destination");
+        body.put("location_type", "start");
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -558,7 +562,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -579,13 +583,13 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM));
 
     }
 
@@ -602,13 +606,13 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM));
     }
 
     @Test
@@ -625,7 +629,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -651,13 +655,13 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
 
     }
 
@@ -675,7 +679,7 @@ class ParamsTest extends ServiceTest {
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
@@ -699,82 +703,38 @@ class ParamsTest extends ServiceTest {
         // ten, 101, -1
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_FORMAT));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_FORMAT));
 
         body.put("smoothing", "-1");
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
 
         body.put("smoothing", "101");
 
         given()
                 .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
+                .pathParam("profile", getParameter("hgvProfile"))
                 .body(body.toString())
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
                 .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
+                .body("error.code", Matchers.is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
 
     }
 
-    @Test
-    void expectUnknownProfile() {
-
-        JSONArray ranges = new JSONArray();
-        ranges.put(600);
-        ranges.put(400);
-        ranges.put(300);
-        JSONObject body = new JSONObject();
-        body.put("locations", getParameter("locations_1"));
-        body.put("range", ranges);
-        body.put("range_type", "time");
-        body.put("interval", getParameter("interval_200"));
-        body.put("location_type", "destination");
-
-        given()
-                .headers(geoJsonContent)
-                .pathParam("profile", "wrongProfile")
-                .body(body.toString())
-                .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
-                .then()
-                .statusCode(400)
-                .body("error.code", is(IsochronesErrorCodes.INVALID_PARAMETER_VALUE));
-
-    }
-
-    @Test
-    void testAcceptsTimeParameter() {
-        JSONArray ranges = new JSONArray();
-        ranges.put(400);
-        JSONObject body = new JSONObject();
-        body.put("locations", getParameter("locations_1"));
-        body.put("range", ranges);
-        body.put("time", "2023-03-20T12:18:00");
-
-        given().log().ifValidationFails()
-                .headers(geoJsonContent)
-                .pathParam("profile", getParameter("carProfile"))
-                .body(body.toString())
-                .when()
-                .post(getEndPointPath() + "/{profile}/geojson")
-                .then().log().ifValidationFails()
-                .statusCode(200);
-    }
 }
