+++ ors-api/src/test/java/org/heigit/ors/apitests/matrix/ResultTest.java
@@ -16,11 +16,9 @@ package org.heigit.ors.apitests.matrix;
 import io.restassured.RestAssured;
 import io.restassured.config.RestAssuredConfig;
 import io.restassured.path.json.config.JsonPathConfig;
-import org.heigit.ors.api.requests.common.APIEnums;
-import org.heigit.ors.api.requests.matrix.MatrixRequest;
+import org.heigit.ors.apitests.common.VersionAnnotation;
 import org.heigit.ors.apitests.common.EndPointAnnotation;
 import org.heigit.ors.apitests.common.ServiceTest;
-import org.heigit.ors.apitests.common.VersionAnnotation;
 import org.heigit.ors.apitests.utils.HelperFunctions;
 import org.json.JSONArray;
 import org.json.JSONObject;
