+++ ors-api/src/test/java/org/heigit/ors/apitests/utils/HelperFunctions.java
